Imported upstream version '3.6.1' of 'upstream'

+++ test/performance/throughput/ThroughputSubscriber.cpp
@@ -781,7 +781,7 @@ void ThroughputSubscriber::run()
             {
                 command_sample.m_totaltime = std::numeric_limits<uint64_t>::min();
             }
-            else if (total_time_count > std::numeric_limits<uint64_t>::max())
+            else if (total_time_count > std::numeric_limits<double>::max())
             {
                 command_sample.m_totaltime = std::numeric_limits<uint64_t>::max();
             }
@@ -797,8 +797,8 @@ void ThroughputSubscriber::run()
                       << (double)(command_sample.m_receivedsamples) * 1000000 /
                 command_sample.m_totaltime
                       << std::endl;
-            std::cout << "Test of size " << command_sample.m_size << " and demand " << command_sample.m_demand <<
-                " ends." << std::endl;
+            std::cout << "Test of size " << command_sample.m_size << " and demand " << command_sample.m_demand
+                      << " ends." << std::endl;
             command_writer_->write(&command_sample);
 
             std::cout << "-----------------------------------------------------------------------" << std::endl;
